requirements.txt
pandas>=2.0.0
numpy>=1.24.0
pytest>=8.0.0
alpaca-py>=0.30.0
python-dotenv>=1.0.0
pytz>=2024.1
